Revert "restore lines of code"
This reverts commit 1673031f41cee867c1864f88d8c7617b9be8332b.

+++ src/client/java/qualet/irlite/client/forms/PointLightFormRenderer.java
@@ -40,7 +40,7 @@ public class PointLightFormRenderer extends AbstractLightFormRenderer<PointLight
         Vector3d p = IRLightPositionResolver.resolve(context);
         Color c = this.form.color.get();
         LightRegistry.registerPoint(
-            p.x, p.y, p.z,
+            (float) p.x, (float) p.y, (float) p.z,
             c.r, c.g, c.b,
             this.form.intensity.get(), this.form.radius.get(),
             this.form.entitiesOnly.get(), this.form.blocksOnly.get(),
